+++ tests/util/tasks_library.py
@@ -1,9 +1,12 @@
 from __future__ import annotations
 
 import pandas as pd
+import datetime as dt
 
 from pydiverse.pipedag import Blob, Table, materialize
 import sqlalchemy as sa
+
+from pydiverse.pipedag.context import ConfigContext
 
 
 @materialize(input_type=pd.DataFrame, version="1.0")
@@ -131,10 +134,17 @@ def simple_dataframes_with_indexes():
 
 
 def _get_df_query():
-    unions = [
-        sa.select([sa.literal(i).label("col1"), sa.literal(str(i)).label("col2")])
-        for i in range(4)
-    ]
+    try:
+        unions = [
+            sa.select([sa.literal(i).label("col1"), sa.literal(str(i)).label("col2")])
+            for i in range(4)
+        ]
+    except sa.exc.ArgumentError:
+        # this works from sqlalchemy 2.0.0 on
+        unions = [
+            sa.select(sa.literal(i).label("col1"), sa.literal(str(i)).label("col2"))
+            for i in range(4)
+        ]
     return unions[0].union_all(*unions[1:])
 
 
@@ -168,10 +178,70 @@ def simple_lazy_table_with_indexes():
     return Table(query, indexes=[["col2"], ["col2", "col1"]])
 
 
-@materialize
-def pd_dataframe(data: dict[str, list]):
-    df = pd.DataFrame(data)
-    return Table(df)
+def _get_df(data: dict[str, list], use_ext_dtype=False, cap_dates=False):
+    data = data.copy()
+    dtypes = {}
+    for col in list(data.keys()):
+        if type(data[col][0]) == dt.date:
+            data[col] = [dt.datetime.combine(d, dt.time()) for d in data[col]]
+        if cap_dates:
+            min_datetime = dt.datetime(1900, 1, 1, 0, 0, 0)
+            max_datetime = dt.datetime(2199, 12, 31, 0, 0, 0)
+            min_date = dt.date(1900, 1, 1)
+            max_date = dt.date(2199, 12, 31)
+            if type(data[col][0]) == dt.date:
+                data[col + "_year"] = [d.year for d in data[col]]
+                dtypes[col + "_year"] = pd.Int16Dtype()
+                data[col] = [max(min(v, max_date), min_date) for v in data[col]]
+                dtypes[col] = "datetime64[ns]"
+            elif type(data[col][0]) == dt.datetime:
+                data[col + "_year"] = [d.year for d in data[col]]
+                dtypes[col + "_year"] = pd.Int16Dtype()
+                data[col] = [max(min(v, max_datetime), min_datetime) for v in data[col]]
+                dtypes[col] = "datetime64[ns]"
+        if use_ext_dtype:
+            if type(data[col][0]) == str:
+                dtypes[col] = pd.StringDtype()
+            elif type(data[col][0]) == int:
+                dtypes[col] = pd.Int64Dtype()
+            elif type(data[col][0]) == bool:
+                dtypes[col] = pd.BooleanDtype()
+    return pd.DataFrame(
+        {
+            col: pd.Series(values, dtype=dtypes[col])
+            if col in dtypes
+            else pd.Series(values)
+            for col, values in data.items()
+        }
+    )
+
+
+@materialize(version="1.0")
+def pd_dataframe(data: dict[str, list], use_ext_dtype=False, cap_dates=False):
+    df = _get_df(data, use_ext_dtype, cap_dates)
+    kwargs = {}
+    if not cap_dates:
+        kwargs["type_map"] = {
+            col: sa.Date for col, items in data.items() if type(items[0]) in [dt.date]
+        }
+        kwargs["type_map"].update(
+            {
+                col: sa.DateTime
+                for col, items in data.items()
+                if type(items[0]) in [dt.datetime]
+            }
+        )
+    return Table(df, **kwargs)
+
+
+@materialize(input_type=pd.DataFrame, version="1.0")
+def pd_dataframe_assert(df_actual: pd.DataFrame, data: dict[str, list]):
+    df_expected = _get_df(data, use_ext_dtype=True, cap_dates=True)
+    if ConfigContext.get().store.table_store.engine.dialect.name == "ibm_db_sa":
+        for col in data:
+            if type(data[col][0]) == bool:
+                df_expected[col] = df_expected[col].astype(pd.Int16Dtype())
+    pd.testing.assert_frame_equal(df_expected, df_actual)
 
 
 @materialize